Split SerializedBaseOperator from serde logic
This allows protected processes to not rely on serde code (which
requires SDK) when using SerializedBaseOperator.

I have not rewritten imports (point to serialized_objects for now)
since those rewrites would conflict pretty badly with my concurrent work
on SerializedDAG. They will be changed later, after the SerializedDAG
split is merged.
